deleted bike app, all of its functionality moved to api/, added selectors and basic services
