+++ Sociale/settings.py
@@ -25,7 +25,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = 'django-insecure-1_6zz+^7dl2oe&-6vipw3-_-g2&m2ux1cw6qb@^^8xk_u8h%*7'
 # CIPHER_KEY = os.getenv("CIPHER_KEY")
-GMAPS_KEY = "AIzaSyA2OtUrY0BGRpWtzLj1Hi2VrKb5YVcGYqs"#os.getenv("MAPS")
+MAPS_KEY = os.getenv("MAPS_KEY")
 
 HOST = os.environ.get('HOST', default="http://localhost:8000/")
 # SECURITY WARNING: don't run with debug turned on in production!
@@ -45,7 +45,6 @@ ALLOWED_HOSTS = ['192.168.1.58','127.0.0.1','localhost']
 # Application definition
 
 INSTALLED_APPS = [
-    'bike_map.apps.BikeMapConfig',
     'base.apps.BaseConfig',
     'django.contrib.admin',
     'django.contrib.auth',
@@ -53,6 +52,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+
+    'rest_framework',
+    'rest_framework_gis',
 ]
 
 MIDDLEWARE = [
@@ -133,7 +135,7 @@ USE_TZ = True
 
 STATIC_URL = 'static/'
 STATICFILES_DIRS = [
-    "Sociale/static",
+    BASE_DIR / 'base' / 'static',
 ]
 
 # Default primary key field type
