--- Sociale/static/base/assets/css/Navbar-With-Button-icons.css
@@ -1,130 +1,0 @@
-.bs-icon {
-  --bs-icon-size: .75rem;
-  display: flex;
-  flex-shrink: 0;
-  justify-content: center;
-  align-items: center;
-  font-size: var(--bs-icon-size);
-  width: calc(var(--bs-icon-size) * 2);
-  height: calc(var(--bs-icon-size) * 2);
-  color: var(--bs-primary);
-}
-
-.bs-icon-xs {
-  --bs-icon-size: 1rem;
-  width: calc(var(--bs-icon-size) * 1.5);
-  height: calc(var(--bs-icon-size) * 1.5);
-}
-
-.bs-icon-sm {
-  --bs-icon-size: 1rem;
-}
-
-.bs-icon-md {
-  --bs-icon-size: 1.5rem;
-}
-
-.bs-icon-lg {
-  --bs-icon-size: 2rem;
-}
-
-.bs-icon-xl {
-  --bs-icon-size: 2.5rem;
-}
-
-.bs-icon.bs-icon-primary {
-  color: var(--bs-white);
-  background: var(--bs-primary);
-}
-
-.bs-icon.bs-icon-primary-light {
-  color: var(--bs-primary);
-  background: rgba(var(--bs-primary-rgb), .2);
-}
-
-.bs-icon.bs-icon-semi-white {
-  color: var(--bs-primary);
-  background: rgba(255, 255, 255, .5);
-}
-
-.bs-icon.bs-icon-rounded {
-  border-radius: .5rem;
-}
-
-.bs-icon.bs-icon-circle {
-  border-radius: 50%;
-}
-
-#hero-paragraph2 {
-}
-
-p {
-  font-size: medium;
-}
-
-#hero-paragraph1 {
-  padding-left: 0px;
-  padding-bottom: 0px;
-  margin-bottom: 14px;
-}
-
-.btn {
-}
-
-#hero-btn:hover {
-  border-color: #ffffff;
-  background: white;
-  color: #FC6701;
-}
-
-#hero-btn {
-  color: rgb(255,255,255);
-  border-color: rgb(255,255,255);
-  background: #21252900;
-  padding-right: 20px;
-}
-
-#about {
-  background: var(--bs-body-bg);
-}
-
-#about-h1 {
-  color: #fc6701;
-  font-size: xx-large;
-  text-align: center;
-  font-family: Inter, sans-serif;
-  letter-spacing: 0.025em;
-  font-weight: bold;
-}
-
-.curve {
-  position: absolute;
-  height: 250px;
-  width: 100%;
-  bottom: 0;
-  text-align: center;
-}
-
-.curve::before {
-  content: '';
-  display: block;
-  position: absolute;
-  border-radius: 100% 50%;
-  width: 55%;
-  height: 100%;
-  transform: translate(85%, 60%);
-  background-color: hsl(216, 21%, 16%);
-}
-
-.curve::after {
-  content: '';
-  display: block;
-  position: absolute;
-  border-radius: 100% 50%;
-  width: 55%;
-  height: 100%;
-  background-color: #3c31dd;
-  transform: translate(-4%, 40%);
-  z-index: -1;
-}
-
