--- Sociale/static/base/assets/css/styles.css
@@ -1,24 +1,0 @@
-body {
-  background: #f3f4f5;
-}
-
-.btn {
-  border-radius: 20px;
-  margin-left: -1px;
-  margin-right: 16px;
-  padding-left: 20px;
-  padding-right: 20px;
-}
-
-#navbar-logo {
-  font-weight: bold;
-  color: #FC6701;
-  font-size: 25px;
-}
-
-.navbar {
-  background: rgb(241,242,244, 0.8);
-  border: 1px solid #dfe3e7 ;
-  padding-left: 0px;
-}
-
