+++ base/templates/base/index.html
@@ -1,7 +1,7 @@
 {% extends "base/core.html" %}
 {% load static %}
 {% block content %}
-<body style="background: linear-gradient(-60deg, #ff5858 0%, #f09819 100%), #ffffff;">
+<body>
 
 <!-- Navbar -->
     <nav class="navbar navbar-expand-md bg-body py-3">
@@ -151,6 +151,7 @@
         </div>
     </section>
     <script src = "{% static 'base/assets/bootstrap/js/bootstrap.min.js' %}" ></script>
+    <!-- delete -->
     <script>
         
         document.getElementById('ContactForm').addEventListener('submit', function(event) {
