+++ base/urls.py
@@ -7,5 +7,7 @@ urlpatterns = [
     path('register/', views.register, name='register'),
     path('logout/',views.logout_view, name='logout'),
     path('password-reset/',views.password_reset_view, name='password-reset'),
-    path('password-reset/<uuid:uuid>/',views.verify_reset_view, name='verify-reset')
+    path('password-reset/<uuid:uuid>/',views.verify_reset_view, name='verify-reset'),
+    path('/map', views.home, name="map-home"),
+
 ]
