--- bike_map/requirements.txt
@@ -1,2 +1,0 @@
-Django
-dotenv
