--- bike_map/templates/bike_map/index.html
@@ -1,27 +1,0 @@
-{% load static %}
-<html>
-  <head>
-    <title>Add a Map using HTML</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css">
-  </head>
-  <body>
-    <div style="position: relative; width: 100%; height: 100vh;text-align:center;">
-      <a href="{% url 'scanner' %}">
-          <button type="button" class="btn btn-primary position-absolute" 
-                  style="z-index: 2; background: #FC6701; border-color: #FC6700;top:92%; ">
-              <i class="bi bi-qr-code"></i> Scan
-          </button>
-      </a>
-      <div id="map" style="height: 100%; z-index: 1;"></div>
-  </div>
-  
-  
-    <!-- TODO -->
-    
-    <script src="{% static 'bike_map/script.js' %}"></script>
-    <script>
-      initMap();
-    </script>
-  </body>
-</html>
