--- bike_map/templates/bike_map/map-index.html
@@ -1,39 +1,0 @@
-{% extends 'base/core.html' %}
-{% load static %}
-{% block content %}
-<body>
-    <div class="row" style="height: 75vh;">
-        <div class="col">
-            <div id="map" style="height: 100%; z-index: 1;"></div>
-        </div>
-    </div>
-    <div class="row" style="height: 27vh;">
-        <div class="col">
-            <div class="rounded-top" style="height: 100%;background: #FFFAFA;width: 100%;box-shadow: 0px 10px rgba(0,0,0,0.2);">
-                <div class="row" style="height: 100%;border-color: var(--bs-emphasis-color);">
-                    <div class="col" style="height: 100%;">
-                        <div class="row" style="height: 50%;">
-                            <div class="col" style="text-align: center; display: inline-block;">
-                                <h1 context-username="{{ user.username }}" id = "menuHeader"class="flex-shrink-1" style="font-weight: bold;margin-top: 0px;margin-bottom: 0px;padding-top: 15px;">/h1>
-                            </div>
-                        </div>
-                        <div class="row" style="height: 50%;">
-                            <div class="col align-middle" style="text-align: center;">
-                                <button id="menuButton" class="btn btn-primary" type="button" style="margin-right: 0px;background: rgb(0,0,0);border-color: rgb(0,0,0);box-shadow: 0px 0px rgb(0,0,0);padding-left: 61px;padding-right: 61px;"><i class="fa-solid fa-qrcode"></i> Scan</button>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <script src="https://use.fontawesome.com/releases/v6.2.0/js/all.js"></script>
-    <script type="text/javascript">
-        const apiKey = "{{ api_key }}";
-    </script>
-    <script src="{% static 'base/script.js' %}"></script>
-    <script>
-    runApp();
-    </script>
-</body>
-{% endblock %}
