--- bike_map/templates/bike_map/scanner.html
@@ -1,24 +1,0 @@
-{% load static %}
-<html>
-  <head>
-    <title>QR</title>
-  </head>
-  <body>
-    <!-- TODO -->
-    <div style="width: 600px" id="reader"></div>
-    <script src="{% static 'bike_map/script.js' %}"></script>
-    <script src="https://unpkg.com/html5-qrcode"></script>
-    <script>
-
-async function onScanSuccess(decodedText, decodedResult) {
-    sessionStorage.setItem('bike_id', decodedText);
-    window.location.href = "{% url 'map-home' %}";
-}
-function onScanError(errorMessage) {
-  // handle on error condition, with error message
-}
-    var html5QrcodeScanner = new Html5QrcodeScanner("reader", { fps: 10, qrbox: 250,videoConstraints: {facingMode: { exact: "environment" },} });
-    html5QrcodeScanner.render(onScanSuccess, onScanError);
-    </script>
-  </body>
-</html>
